adding tests and refactoring folder structure
