+++ __tests__/app.test.js
@@ -43,13 +43,23 @@ describe("GET /", () => {
   });
 });
 
-describe('Blog Routes', () => {
+describe("Blog Routes", () => {
   beforeAll(async () => {
     // Sync the database and create some test data
     await BlogPost.sync({ force: true });
     await BlogPost.bulkCreate([
-      { title: 'new', author: 'new', content: 'Content of new post', created_at: new Date('2023-01-02') },
-      { title: 'old', author: 'old', content: 'Content of old post', created_at: new Date('2023-01-01') }
+      {
+        title: "new",
+        author: "new",
+        content: "Content of new post",
+        created_at: new Date("2023-01-02"),
+      },
+      {
+        title: "old",
+        author: "old",
+        content: "Content of old post",
+        created_at: new Date("2023-01-01"),
+      },
     ]);
   });
 
@@ -58,95 +68,135 @@ describe('Blog Routes', () => {
     await BlogPost.destroy({ where: {}, truncate: true });
   });
 
-  describe('GET /', () => {
-    it('should return all posts sorted by created_at DESC by default', async () => {
-      const response = await request(app).get('/');
+  describe("GET /", () => {
+    it("should return all posts sorted by created_at DESC by default", async () => {
+      const response = await request(app).get("/");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('new by new');
-      expect(response.text).toContain('old by old');
-      expect(response.text.indexOf('new by new')).toBeLessThan(response.text.indexOf('old by old'));
+      expect(response.text).toContain("new by new");
+      expect(response.text).toContain("old by old");
+      expect(response.text.indexOf("new by new")).toBeLessThan(
+        response.text.indexOf("old by old")
+      );
     });
 
-    it('should return posts sorted by created_at ASC', async () => {
-      const response = await request(app).get('/?q=&sort=created_at%3Aasc');
+    it("should return posts sorted by created_at ASC", async () => {
+      const response = await request(app).get("/?q=&sort=created_at%3Aasc");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('new by new');
-      expect(response.text).toContain('old by old');
-      expect(response.text.indexOf('old by old')).toBeLessThan(response.text.indexOf('new by new'));
+      expect(response.text).toContain("new by new");
+      expect(response.text).toContain("old by old");
+      expect(response.text.indexOf("old by old")).toBeLessThan(
+        response.text.indexOf("new by new")
+      );
     });
 
-    it('should return all posts sorted by A-Z', async () => {
-      const response = await request(app).get('/?q=&sort=title%3Aasc');
+    it("should return all posts sorted by A-Z", async () => {
+      const response = await request(app).get("/?q=&sort=title%3Aasc");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('new by new');
-      expect(response.text).toContain('old by old');
-      expect(response.text.indexOf('new by new')).toBeLessThan(response.text.indexOf('old by old'));
+      expect(response.text).toContain("new by new");
+      expect(response.text).toContain("old by old");
+      expect(response.text.indexOf("new by new")).toBeLessThan(
+        response.text.indexOf("old by old")
+      );
     });
 
-    it('should return all posts sorted by Z-A', async () => {
-      const response = await request(app).get('/?q=&sort=title%3Adesc');
+    it("should return all posts sorted by Z-A", async () => {
+      const response = await request(app).get("/?q=&sort=title%3Adesc");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('new by new');
-      expect(response.text).toContain('old by old');
-      expect(response.text.indexOf('old by old')).toBeLessThan(response.text.indexOf('new by new'));
+      expect(response.text).toContain("new by new");
+      expect(response.text).toContain("old by old");
+      expect(response.text.indexOf("old by old")).toBeLessThan(
+        response.text.indexOf("new by new")
+      );
     });
 
-    it('should return posts matching the search query', async () => {
-      const response = await request(app).get('/?q=old');
+    it("should return posts matching the search query", async () => {
+      const response = await request(app).get("/?q=old");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('old by old');
-      expect(response.text).not.toContain('new by new');
+      expect(response.text).toContain("old by old");
+      expect(response.text).not.toContain("new by new");
     });
 
-    it('should return posts matching the search query case-insensitive', async () => {
-      const response = await request(app).get('/?q=NEW');
+    it("should return posts matching the search query case-insensitive", async () => {
+      const response = await request(app).get("/?q=NEW");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('new by new');
-      expect(response.text).not.toContain('old by old');
+      expect(response.text).toContain("new by new");
+      expect(response.text).not.toContain("old by old");
     });
 
-    it('should return all posts sorted by by created_at ASC and matching the search query', async () => {
+    it("should return all posts sorted by by created_at ASC and matching the search query", async () => {
       await BlogPost.bulkCreate([
-        { title: 'old2', author: 'old', content: 'Content of new post', created_at: new Date('2023-01-02') },
-        { title: 'old3', author: 'old', content: 'Content of newer post', created_at: new Date('2023-01-03') }
+        {
+          title: "old2",
+          author: "old",
+          content: "Content of new post",
+          created_at: new Date("2023-01-02"),
+        },
+        {
+          title: "old3",
+          author: "old",
+          content: "Content of newer post",
+          created_at: new Date("2023-01-03"),
+        },
       ]);
-      const response = await request(app).get('/?q=old&sort=created_at%3Aasc');
+      const response = await request(app).get("/?q=old&sort=created_at%3Aasc");
 
       expect(response.status).toBe(200);
-      expect(response.text).toContain('old by old');
-      expect(response.text).toContain('old2 by old');
-      expect(response.text).toContain('old3 by old');
-      expect(response.text.indexOf('old by old')).toBeLessThan(response.text.indexOf('old2 by old'));
-      expect(response.text.indexOf('old by old')).toBeLessThan(response.text.indexOf('old3 by old'));
-      expect(response.text.indexOf('old2 by old')).toBeLessThan(response.text.indexOf('old3 by old'));
+      expect(response.text).toContain("old by old");
+      expect(response.text).toContain("old2 by old");
+      expect(response.text).toContain("old3 by old");
+      expect(response.text.indexOf("old by old")).toBeLessThan(
+        response.text.indexOf("old2 by old")
+      );
+      expect(response.text.indexOf("old by old")).toBeLessThan(
+        response.text.indexOf("old3 by old")
+      );
+      expect(response.text.indexOf("old2 by old")).toBeLessThan(
+        response.text.indexOf("old3 by old")
+      );
     });
 
-    it('should return all posts sorted by by created_at DESC and matching the search query', async () => {
+    it("should return all posts sorted by by created_at DESC and matching the search query", async () => {
       await BlogPost.bulkCreate([
-        { title: 'old2', author: 'old', content: 'Content of new post', created_at: new Date('2023-01-02') },
-        { title: 'old3', author: 'old', content: 'Content of newer post', created_at: new Date('2023-01-03') }
+        {
+          title: "old2",
+          author: "old",
+          content: "Content of new post",
+          created_at: new Date("2023-01-02"),
+        },
+        {
+          title: "old3",
+          author: "old",
+          content: "Content of newer post",
+          created_at: new Date("2023-01-03"),
+        },
       ]);
-      const response = await request(app).get('/?q=old&sort=created_at%3Adesc');
+      const response = await request(app).get("/?q=old&sort=created_at%3Adesc");
 
       expect(response.status).toBe(200);
-      expect(response.text).toContain('old by old');
-      expect(response.text).toContain('old2 by old');
-      expect(response.text).toContain('old3 by old');
-      expect(response.text.indexOf('old3 by old')).toBeLessThan(response.text.indexOf('old2 by old'));
-      expect(response.text.indexOf('old3 by old')).toBeLessThan(response.text.indexOf('old by old'));
-      expect(response.text.indexOf('old2 by old')).toBeLessThan(response.text.indexOf('old by old'));
+      expect(response.text).toContain("old by old");
+      expect(response.text).toContain("old2 by old");
+      expect(response.text).toContain("old3 by old");
+      expect(response.text.indexOf("old3 by old")).toBeLessThan(
+        response.text.indexOf("old2 by old")
+      );
+      expect(response.text.indexOf("old3 by old")).toBeLessThan(
+        response.text.indexOf("old by old")
+      );
+      expect(response.text.indexOf("old2 by old")).toBeLessThan(
+        response.text.indexOf("old by old")
+      );
     });
 
-    it('should return a message when no posts match the search query', async () => {
-      const response = await request(app).get('/?q=nonexistent');
+    it("should return a message when no posts match the search query", async () => {
+      const response = await request(app).get("/?q=nonexistent");
       expect(response.status).toBe(200);
-      expect(response.text).toContain('Sorry, no posts found.');
-      expect(response.text).not.toContain('old by old');
-      expect(response.text).not.toContain('new by new');
+      expect(response.text).toContain("Sorry, no posts found.");
+      expect(response.text).not.toContain("old by old");
+      expect(response.text).not.toContain("new by new");
     });
 
-    it('should return 404 if post not found', async () => {
-      const response = await request(app).get('/post/999');
+    it("should return 404 if post not found", async () => {
+      const response = await request(app).get("/post/999");
       expect(response.status).toBe(404);
     });
   });
